add api hoa don

+++ BE/src/main/java/com/example/datn_f5_store/dto/VoucherDto.java
@@ -23,7 +23,7 @@ public class VoucherDto {
 
     private String ten;
 
-    private Integer giaTriVoucher;
+    private Double giaTriVoucher;
 
     private String kieuGiamGia;
